updaed G1 form delete  , to cehck with Form F3

+++ RAMS/Web/RAMMS.Business.ServiceProvider/Services/FormG1G2Service.cs
@@ -222,9 +222,13 @@ namespace RAMMS.Business.ServiceProvider.Services
         }
         public int Delete(int id)
         {
-            if (id > 0)
+            if (id > 0 && !_repo.isF3Exist(id))
             {
                 id = _repo.DeleteHeader(new RmFormG1Hdr() { Fg1hActiveYn = false, Fg1hPkRefNo = id });
+            }
+            else
+            {
+                return -1;
             }
             return id;
         }
